Done Sprint 1 (finishing fe soon)

+++ database/migrations/2025_03_25_151242_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('productName');
             $table->json('productPhoto');
             $table->integer('productPrice');
-            $table->float('productNetto');
+            $table->integer('productNetto');
             $table->string('productUnit');
             $table->text('productDescription');
             $table->enum('productType',['setengah jadi','siap pakai']);
